// src/app/core/types/enum.ts
export enum StorageEnum {
  BEGIN = "begin",
  AUTH = "auth",
  WORKSPACE = "workspace",
  MIGRATION = "migrations",
  STORE = "stores",
  STORE_ROUTE = "store_route",
  TIPO_PEDIDO = "tipo_pedido",
  PEDIDO_TEMP = "pedido_temp",
  PRAZO_PAGAMENTO = "prazo_pagamento",
  DISTRIBUIDORES = "distribuidores",
  CONDICAO_DISTRIBUIDOR_POR_PEDIDO_ESPECIAL = "condicao_distribuidor_por_pedido_especial",
  TABLOIDE = "tabloide",
  CONDICAO_COMERCIAL = "condicao_comercial",
  PRAZO_PAGAMENTO_LISTA = "prazo_pagamento_lista",
  DISTRIBUIDORES_PRAZO_PAGAMENTO_LISTA = "distribuidores_prazo_pagamento_lista",
  CHAVE_UNICA_PEDIDO = "chave_unica_pedido",
  PARAMETRIZACAO_LOOPING = "parametrizacao_looping",
  PRODUTOS_FINAL = "produtos_final",
  DATAS_PROGRAMADAS_FINAL = "datas_programadas_final",
  NUMERO_CLIENTE_FINAL = "numero_cliente_final",
  OBSERVACAO = 'observacao',
  PRODUTOS_DETALHE = "produtos_detalhe",
  SKU = "sku",
  STORE_DETAILS = "store_details",
  SYNC = "sync",
  VISITA_CHECKIN_ATIVO = "visita_checkin_ativo",
  VISITA_TEMP = "visita_temp",
  CURRENT_DATE = "current_date",
  ORDER_EDIT = "order_edit",
  COMBOS_FINAL = "combos_final",
  LAST_SYNC_STORE = "last_sync_store",
  LAST_SYNC_STORE_ROUTE = "last_sync_store_route",
  LAST_SYNC_INFO_REQUEST = "last_sync_info_request",
  LAST_SYNC_INFO_REQUEST_SPECIAL = "last_sync_info_request_special",
  LAST_SYNC_INFO_REQUEST_DEFAULT = "last_sync_info_request_default",
  LAST_SYNC_INFO_REQUEST_REP = "last_sync_info_request_rep",
  LAST_SYNC_REQUESTS_SYNC = "last_sync_requests_sync",
  LAST_SYNC_THEME = "last_sync_theme",
  LAST_SYNC_INFO_VISIT = "last_sync_info_visit",
  LAST_SYNC_VISIT = "last_sync_visit",
  LAST_SYNC_CONTENT = "last_sync_content",
  LAST_SYNC_RESEARCH_INDUSTRY_PRODUCT = "last_sync_research_industry_product",
  LAST_SYNC_RESEARCH_INDUSTRY_MERCHAN = "last_sync_research_industry_merchan",
  LAST_SYNC_RESEARCH_SHARE_OF_SHELF = "last_sync_research_share_of_shelf",
  LAST_SYNC_RESEARCH_CONCURRENT_MERCHAN = "last_sync_research_concurrent_merchan",
  LAST_SYNC_RESEARCH_CONCURRENT_PRODUCT = "last_sync_research_concurrent_product",
  LAST_SYNC_RESEARCH_COMPLEMENTARY = "last_sync_research_complementary",
  LAST_SYNC_RESEARCH_TRADE = "last_sync_research_trade",
  FCM_TOKEN = "fcm_token",
  STORE_REQUESTEDS_ON_DAY = "store_requesteds_on_day",
  PAGE_NAME_REDIRECT = "page_name_redirect",
  PARAMETRIZACAO_TEMP = "parametrizacao_temp",
  PARAMETRIZACAO_CHECKIN = "HabilitaCheckInCheckOut",
  PARAMETRIZACAO_DISTANCIACHECKIN = "ParametrizacaoDistanciaCheckIn",

  /** Mix Ideal */
  PRODUTOS_SELECIONADOS_MIX_IDEAL = "produtos_selecionados_mix_ideal",
  PRODUTOS_MIX_IDEAL_REVALIDADOS = "produtos_mix_ideal_revalidados",
  PRODUTOS_ENVIADOS_VALIDACAO_MIX_IDEAL = "produtos_enviados_validacao_mix_ideal",
  PRODUTOS_MIX_IDEAL_VALIDOS = "produtos_mix_ideal_validos",
  LISTA_MIX_IDEAL_POR_LOJA = "lista_mix_ideal_por_loja",
  MIX_IDEAL_ALREADY_UPDATE_ITEMS = "mix_ideal_already_update_items",
  MIX_VIEW = "mix_ideal_page_already_visited"
}

export enum StatusSyncronization {
  SENT = 1,
  NOT_SENT = 2,
  RECEIVE_WITH_ERROR = 3,
  RECEIVE_WITH_SUCCESS = 4,
  SENT_FOR_APPROVAL = 5,
  RECEIVE_REPROVED = 6,
  RECEIVE_APPROVED = 7,
  RECEIVE_AUTH_ERROR = 8
}